requirements.txt
protobuf
sentencepiece
torchaudio
transformers

datasets
librosa
soundfile

